+++ boards/posix/nrf52_bsim/Kconfig.board
@@ -7,7 +7,6 @@ config BOARD_NRF52_BSIM
 	select SOC_COMPATIBLE_NRF
 	select SOC_COMPATIBLE_NRF52X
 	select SOC_COMPATIBLE_NRF52833
-	select CONSOLE_HAS_DRIVER
 	select NRF_RTC_TIMER
 	select CLOCK_CONTROL
 	select HAS_NRFX
