+++ drivers/console/CMakeLists.txt
@@ -1,6 +1,12 @@
 # SPDX-License-Identifier: Apache-2.0
 
 zephyr_library()
+
+if (${CONFIG_BSIM_CONSOLE})
+  zephyr_library_sources(bsim_console.c)
+  zephyr_library_include_directories($ENV{BSIM_COMPONENTS_PATH}/libUtilv1/src/)
+endif()
+
 zephyr_library_sources_ifdef(CONFIG_GSM_MUX gsm_mux.c)
 zephyr_library_sources_ifdef(CONFIG_IPM_CONSOLE_RECEIVER ipm_console_receiver.c)
 zephyr_library_sources_ifdef(CONFIG_IPM_CONSOLE_SENDER ipm_console_sender.c)
