+++ drivers/console/Kconfig
@@ -227,6 +227,20 @@ config XTENSA_SIM_CONSOLE
 	help
 	  Use simulator console to print messages.
 
+config BSIM_CONSOLE
+	bool "Use the BabbleSim tracing system for console output"
+	depends on ARCH_POSIX
+	select CONSOLE_HAS_DRIVER
+	help
+	  Use the BabbleSim tracing system for the Zephyr console.
+
+config BSIM_CONSOLE_INIT_PRIORITY
+	int "Init priority"
+	default 99
+	depends on BSIM_CONSOLE
+	help
+	  Device driver initialization priority.
+
 config NATIVE_POSIX_CONSOLE
 	bool "Use the host terminal for console"
 	depends on ARCH_POSIX
